mysite/requirements.txt
Django==3.2
djangorestframework==3.12

accelerate==0.24.1
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
amqp==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asgiref==3.7.2
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
audioread==3.0.1
billiard==4.2.0
blinker==1.6.3
cachetools==5.3.1
celery==5.3.5
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
ci-sdr==0.0.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7
ctc-segmentation==1.7.4
Cython==3.0.3
datasets==2.14.6
decorator==5.1.1
dill==0.3.7
Distance==0.1.3
Django==4.2.6
django-extensions==3.2.3
editdistance==0.6.2
einops==0.7.0
espnet==202308
espnet-model-zoo==0.1.7
espnet-tts-frontend==0.0.3
exceptiongroup==1.1.3
fast-bss-eval==0.1.3
fastapi==0.103.2
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.9.2
g2p-en==2.1.0
gitdb==4.0.10
GitPython==3.1.37
gunicorn==21.2.0
h11==0.14.0
h5py==3.10.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
huggingface-hub==0.17.3
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
importlib-metadata==4.13.0
inflect==7.0.0
jaconv==0.3.4
jamo==0.4.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kaldiio==2.18.0
kombu==5.3.3
librosa==0.9.2
llvmlite==0.41.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.1
nltk==3.8.1
numba==0.58.0
numpy==1.23.5
omegaconf==2.3.0
opt-einsum==3.3.0
packaging==23.2
pandas==2.0.3
Pillow==10.0.1
platformdirs==3.11.0
pooch==1.7.0
prompt-toolkit==3.0.41
protobuf==3.20.1
psutil==5.9.6
pyarrow==13.0.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
pylzma==0.5.0
pyopenjtalk==0.3.2
pyopenjtalk-prebuilt==0.3.0
pypinyin==0.44.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytorch-wpe==0.0.1
pytz==2023.3.post1
pyworld==0.3.4
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
resampy==0.4.2
rich==13.6.0
rpds-py==0.10.4
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.10.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soundfile==0.12.1
sqlparse==0.4.4
starlette==0.27.0
streamlit==1.27.2
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.14.1
toml==0.10.2
toolz==0.12.0
torch==2.1.0
torch-complex==0.4.3
torchaudio==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
transformers @ git+https://github.com/huggingface/transformers.git@4309abedbc3b767303a07133e00af95876c4ad0b
typeguard==2.13.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
Unidecode==1.3.7
urllib3==2.0.6
uvicorn==0.23.2
uvloop==0.17.0
validators==0.22.0
vine==5.1.0
watchfiles==0.20.0
wcwidth==0.2.10
websockets==11.0.3
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
django-environ
httpx
python-dotenv
